Added support for Postgre DB/Live Profile
Have created the new live application properites which will connect to a
postgre database.  This is to ensure taht live versions of this
application are hosting their data externally.  Additionally have
excluded the Bootstrap class from running with the live profile as data
will be persistable and this will result in constraint exceptions.

Have also switched the genereated ids to the more efficient 'sequence'
strategy as this allows multi db platform support and better creation
statements.

This now means the application can be run with a live profile by adding
the -Dspring.profiles.active=live parameter to the vm arguments.  If
this is not provided the db will default to the H2 in memory database.