Changed Structure of Calls
This change introduces the TwilioNumbers table and the full apis to
access this.  This allows the application to maintain a list of twilio
numbers and whether they are in use.  Each event will be associated to a
twilio number, if it is in use the event will be stopped by the service.

Additionally, the InboundSms table has been updated to have  status
associated to each message.  This will determine whether a reporter has
responded or not to an outbound message.

Still need to implement the logic using the new structures and implement
unit/integration tests.