Made Structural Changes for Single Message table
As part of the refactor to how we handle incoming message, the
InboundSms table has been merged with the OutboundSms table.  This means
that we have one table BcpEventSms table which logs outbound and inbound
messages.

This means when an event is tarted, each recipient is logged with an SMS
status of sent.

An upcoming change will handle how incoming messages will update these
tables to sms status of received and relevant recipient information.