Added Jms Messaging and Close Event Handling
The following change addresses the issue when an SMS is received to
check if the event should be closed/the twilio number freed up.  This is
done when receivng a message, checking the corresponding event to see if
any other messages in this event have the status sent.  If so, we set
the event as inactive and make the TwilioNumber available again.

This is handled via Jms Messages on the default Active MQ queue.  The
theory here is that this check should not interrupt the process of
responding the incoming SMS and so should be handled separately by the
application.  This also opens the option for the close event process to
contain more functionality e.g. push messages.

This change also introduces some minor name changes and storage issues.