Add soft delete for Twilio Numbers
Previously, the user was able to delete numbers from the database from
the application.  This meant that any history of them would be deleted
as well.  This could aversely affect stats.

This change introduces the 'active' column to the numbers table which
will be set to true by default when creating a number.  When a user
deletes a number this is then set to false rather than removing the
entry ensuring that it is hidden from the application.  Should the user
create the same number again, this entry is simply set to true again,
preserving its previous history.